rps/app/requirements.txt
Flask
uwsgi
python-dotenv